% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-wincred.R
\name{b_wincred_decode}
\alias{b_wincred_decode}
\title{Decode a raw password obtained by b_wincred_get_raw}
\usage{
b_wincred_decode(password, encoding = "auto")
}
\arguments{
\item{password}{A raw byte string returned from \code{b_wincred_get_raw}.}

\item{encoding}{A character value, specifying an encoding to use.
Defaults to 'auto', which will decode either of UTF-8 or UTF-16LE.}
}
\value{
A character value containing a password.
}
\description{
Defaults to 'auto' encoding, which uses \code{b_wincred_decode_auto} to
accomplish the decoding (this works with decoding either UTF-8 or
UTF-16LE encodings). In the case where an encoding is specified,
use that to convert the raw password.
}
\keyword{internal}
