% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-wincred.R
\name{b_wincred_set_with_raw_value}
\alias{b_wincred_set_with_raw_value}
\title{Set a key on a Wincred keyring}
\usage{
b_wincred_set_with_raw_value(self, private, service, username, password,
  keyring)
}
\arguments{
\item{service}{Service name. Must not be empty.}

\item{username}{Username. Might be empty.}

\item{password}{The key text to store.
\enumerate{
\item Check if we are using the default keyring.
\item If yes, then just store the key and we are done.
\item Otherwise check if keyring exists.
\item If not, error and finish.
\item If yes, check if it is locked.
\item If yes, unlock it.
\item Encrypt the key with the AES key, and store it.
}}
}
\description{
Set a key on a Wincred keyring
}
\keyword{internal}
