% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-wincred.R
\name{backend_wincred}
\alias{backend_wincred}
\title{Windows Credential Store keyring backend}
\description{
This backend is the default on Windows. It uses the native Windows
Credential API, and needs at least Windows XP to run.
}
\details{
This backend supports multiple keyrings. Note that multiple keyrings
are implemented in the \code{keyring} R package, using some dummy keyring
keys that represent keyrings and their locked/unlocked state.

See \link{backend} for the documentation of the individual methods.
}
\examples{
\dontrun{
## This only works on Windows
kb <- backend_wincred$new()
kb$keyring_create("foobar")
kb$set_default_keyring("foobar")
kb$set_with_value("service", password = "secret")
kb$get("service")
kb$delete("service")
kb$delete_keyring("foobar")
}
}
\seealso{
Other keyring backends: 
\code{\link{backend_env}},
\code{\link{backend_file}},
\code{\link{backend_macos}},
\code{\link{backend_ppm}},
\code{\link{backend_secret_service}}
}
\concept{keyring backends}
