% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_wincred.R
\name{backend_wincred_get}
\alias{backend_wincred_get}
\title{Get a key from a Wincred keyring}
\usage{
backend_wincred_get(backend, service, username)
}
\arguments{
\item{backend}{Backend object.}

\item{service}{Service name. Must not be empty.}

\item{username}{Username. Might be empty.
\enumerate{
\item We check if the key is on the default keyring.
\item If yes, we just return it.
\item Otherwise check if the keyring is locked.
\item If locked, then unlock it.
\item Get the AES key from the keyring.
\item Decrypt the key with the AES key.
}}
}
\description{
Get a key from a Wincred keyring
}
\keyword{internal}
