% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_backend.R
\name{keyring backends}
\alias{keyring backends}
\alias{default_backend}
\title{Select the default backend for a platform}
\usage{
default_backend()
}
\description{
\enumerate{
\item On Windows, the Windows Credential Store is used.
\item On macOS, Keychain services are selected.
\item Linux uses the Secret Service API.
\item On other operating systems, secrets are stored in environment
variables.
}
}
