% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ keyring }
\alias{ keyring }
\alias{ keyring }
\title{ Access the System Credential Store from R }
\description{
Platform independent API to many system credential store implementations.
  Currently supported: Keychain on 'macOS', Credential Store on 'Windows', the
  Secret Service API and the Gnome keyring API on 'Linux'.
}
