% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-awssecretsmanager.R
\name{backend_awssecretsmanager}
\alias{backend_awssecretsmanager}
\title{AWS Secrets Manager keyring backend}
\description{
This backend must be selected explicitly. It makes calls to the AWS
secretsmanager service.
}
\details{
This backend does not support keyrings or user names.  The call to the
AWS service is authenticated by either the user's ceedentials or the IAM
user associated with the process, for example in a docker container.

See \link{backend} for the documentation of the individual methods.

\if{html}{\out{<div class="sourceCode">}}\preformatted{is_available = function(report_error = FALSE)
}\if{html}{\out{</div>}}

Argument:
\itemize{
\item \code{report_error} Whether to throw an error if the Secret Service is
not available.
}
}
\examples{
\dontrun{
##
kb <- backend_awssecretsmanager$new()
kb$set_with_value("service", password = "secret")
kb$get("service")
kb$delete("service")
}
}
\seealso{
Other keyring backends: 
\code{\link{backend_env}},
\code{\link{backend_file}},
\code{\link{backend_macos}},
\code{\link{backend_secret_service}},
\code{\link{backend_wincred}}
}
\concept{keyring backends}
