% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_env.R
\name{backend_env}
\alias{backend_env}
\title{Store secrets in environment variables}
\usage{
backend_env()
}
\description{
This is a simple keyring backend, that stores/uses secrets in
environment variables of the R session.
}
\details{
It does not support multiple keyrings. It also does not support listing
all keys, since there is no way to distinguish between keys and regular
environment variables.

It does support service names and usernames: they will be separated
with a \code{:} character in the name of the environment variable.
}
\examples{
\dontrun{
env <- backend_env()
key_set("r-keyring-test", username = "donaldduck", backend = env)
key_get("r-keyring-test", username = "donaldduck", backend = env)
Sys.getenv("r-keyring-test:donaldduck")

# This is an error
key_list(backend = env)

# Clean up
key_delete("r-keyring-test", username = "donaldduck", backend = env)
}
}
\seealso{
Other keyring backends: \code{\link{backend_macos}},
  \code{\link{backend_secret_service}},
  \code{\link{backend_wincred}}
}
