% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-file.R
\docType{data}
\name{backend_file}
\alias{backend_file}
\title{Store secrets in encrypted files}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
backend_file
}
\description{
This is a simple keyring backend, that stores/uses secrets in encrypted
files.
}
\details{
It supports multiple keyrings.

See \link{backend} for the documentation of the individual methods.
}
\examples{
\dontrun{
kb <- backend_file$new()
}
}
\seealso{
Other keyring backends: \code{\link{backend_env}},
  \code{\link{backend_macos}},
  \code{\link{backend_secret_service}},
  \code{\link{backend_wincred}}
}
\keyword{datasets}
