% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_macos.R
\name{backend_macos}
\alias{backend_macos}
\title{Create a macOS Keychain backend}
\usage{
backend_macos(keyring = NULL)
}
\arguments{
\item{keyring}{Name of the keyring to use. \code{NULL} specifies the
default keyring.}
}
\value{
A backend object that can be used in \code{keyring} functions.
}
\description{
This backend is the default on macOS. It uses the macOS native Keychain
Service API.
}
\details{
It supports multiple keyrings.
}
\seealso{
Other keyring backends: \code{\link{backend_env}},
  \code{\link{backend_secret_service}},
  \code{\link{backend_wincred}}
}
