% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-macos.R
\docType{data}
\name{backend_macos}
\alias{backend_macos}
\title{Create a macOS Keychain backend}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
backend_macos
}
\arguments{
\item{keyring}{Name of the keyring to use. \code{NULL} specifies the
default keyring.}
}
\value{
A backend object that can be used in \code{keyring} functions.
}
\description{
This backend is the default on macOS. It uses the macOS native Keychain
Service API.
}
\details{
It supports multiple keyrings.
}
\seealso{
Other keyring backends: \code{\link{backend_env}},
  \code{\link{backend_secret_service}},
  \code{\link{backend_wincred}}
}
\keyword{datasets}
