% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_secret_service.R
\name{backend_secret_service}
\alias{backend_secret_service}
\title{Create a Secret Service keyring backend}
\usage{
backend_secret_service(keyring = NULL)
}
\arguments{
\item{keyring}{Name of the keyring to use. \code{NULL} specifies the
default keyring.}
}
\value{
A backend object that can be used in \code{keyring} functions.
}
\description{
This backend is the default on Linux. It uses the libsecret library,
and needs a secret service daemon running (e.g. Gnome Keyring, or
KWallet). It uses DBUS to communicate with the secret service daemon.
}
\details{
This backend supports multiple keyrings
}
\seealso{
Other keyring backends: \code{\link{backend_env}},
  \code{\link{backend_macos}},
  \code{\link{backend_wincred}}
}
