% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_wincred.R
\name{backend_wincred}
\alias{backend_wincred}
\title{Create a Windows Credential Store keyring backend}
\usage{
backend_wincred(keyring = NULL)
}
\arguments{
\item{keyring}{Name of the keyring to use. \code{NULL} specifies the
default keyring.}
}
\value{
A backend object that can be used in \code{keyring} functions.
}
\description{
This backend is the default on Windows. It uses the native Windows
Credential API, and needs at least Windows XP to run.
}
\details{
This backend supports multiple keyrings. Note that multiple keyrings
are implemented in the \code{keyring} R package, using some dummy keyring
keys that represent keyrings and their locked/unlocked state.
}
\seealso{
Other keyring backends: \code{\link{backend_env}},
  \code{\link{backend_macos}},
  \code{\link{backend_secret_service}}
}
