% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{key_get}
\alias{key_get}
\alias{key_set}
\alias{key_set_with_value}
\alias{key_delete}
\alias{key_list}
\title{Query, set, delete list keys in a keyring}
\usage{
key_get(service, username = NULL, backend = default_backend())

key_set(service, username = NULL, backend = default_backend())

key_set_with_value(service, username = NULL, password = NULL,
  backend = default_backend())

key_delete(service, username = NULL, backend = default_backend())

key_list(service = NULL, backend = default_backend())
}
\arguments{
\item{service}{Service name, a character scalar.}

\item{username}{Username, a character scalar, or \code{NULL} if the key
is not associated with a username.}

\item{backend}{Backend to use. See \link{backends}.}

\item{password}{The secret to store. For \code{key_set}, it is read from
the console, interactively. \code{key_set_with_value} can be also used
in non-interactive mode.}
}
\value{
\code{key_get} returns a character scalar, the password or other
confidential information that was stored in the key.
\code{key_list} returns a list of keys, i.e. service names and usernames,
in a data frame.
}
\description{
Query, set, delete list keys in a keyring
}
\examples{
# TODO
}
