% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ keyring }
\alias{ keyring }
\alias{ keyring }
\title{ Access the System Keyring from R }
\description{
Modern OSes store passwords, and other confidential information in a keyring with a master password. This package allows access to the keyring of the OS from R. It supports Windows, Mac OS X and Linux systems.
}
